Prioritize VMware
Vagrant's VMware - despite being sold for 79 bux a pop - truly suck in so many ways the devil is laughing. The community is more or less asking HashiCorp in various threads on GitHub if they can be kind enough to officially drop the provider or make it open source. This puts me in a moral dilemma. VMware has a noticeably faster performance than VirtualBox. For a box with a desktop, this is significant. But, Vagrant + VMware = disaster. The provider is completely broken on Vagrant 2.0.3, a note that has gone to GitHub and the readme file on martinanderssondotcom/box-ubuntu-budgie-18-x64.

So what do you do? I have to, for the moment, assume two things: 1) Maybe the VMware provider plugin will improve over time. 2) If user has the VMware provider plugin installed, then he has already "made a choice" to use this guy before VirtualBox.